Removing unnecessary default config and volumes
Mount points are created in the image, but volumes are not automatically
created. The user can still create volumes for these mount points, but is
not required to.

The default configuration file is now copied from /etc/mosquitto/mosquitto.conf
to /mosquitto/config/mosquitto.conf. The user can override this file by
mounting a local configuration file to /mosquitto/config.