Add PKCS11 and file URI support to Mosquitto
This changes allows certs and CAs to be in the form of file URIs and private keys to be either file URIs or PKCS11 compliant URIs.
The private keys can either be stored on the filesystem or in a PKCS11 compliant Trusted Platform Module/Hardware Security Module
(TPM/HSM). This provides more security to the private keys as they cannot be directly copied off of the filesystem.
Changes:
1. Added new API mosquitto_tls_set_uri to accept cert and CA as file URI (RFC8089) and private key as PKCS11 compliant (RFC7512) URI.
2. The above function also accepts paths to the libp11 library and the PKCS11 provider library, which is required
   to access keys stored on a TPM/HSM
3. Modified internal tls_set function to parse URI
4. Modified internal ssl connect function to initialize and call appropriate OpenSSL engine functions to load private key
   from the HSM/TPM
5. Added options for mosquitto_sub and mosquitto_pub to accept URIs and lib paths
6. Added unit tests for the new mosquitto_tls_set_uri function

The above changes were tested using SoftHSM2 as the keystore.

Signed-off-by: Varun Reddy <vareddy@amazon.com>