Prolongate timeout on PUBLISH only when not already waiting for PINGRESP
When a PINGREQ has been sent the PINGRESP is expected within
next keepalive interval and the select timeout is set accordingly.
Sending PUBLISH on the socket may still be possible at this time
although the connection is no longer working. Normally the next
timeout is set to PUBLISH time + another keepalive interval,
but in case the PINGRESP has not arrived yet, it should be kept
to detect a connection problem as soon as possible.

Change-Id: I9c9cd62a228d2f3edb0aba513cd454845bd654d8
Signed-off-by: Sarah Nieswand <Sarah.Nieswand@partner.bmw.de>