Fix mosquitto_loop_stop() hangs on Windows.
Before this commit, mosquitto_loop_forever() would not always
terminate on Windows when calling mosquitto_loop_stop() with
the force flag set to true.

This commit adds explicit thread cancellation points in
mosquitto_loop_forever() as described in issue #117 to
ensure that mosquitto_loop_stop() does not hang when
joining the Mosquitto thread.

Closes #117.

Signed-off-by: Sigmund Vik <sigmund_vik@yahoo.com>