Allow resolving hostnames in mosquitto connect
Allow resolving hostnames before invoking mosquitto connect and
pass hostname and address with mosquitto_add_hostaddress instead.
This prevents that the internal dns resolving using getaddressinfo
blocks the application for a few seconds, if no internet connection
is available.

Note: A hostname is always required, if tls is enabled to verify
the tls hostname. Therefore one can not just replace hostname by
hostaddress during a call to mosquitto_connect.

Providing a hostaddress allows dns resolving with custom dns apis,
which allows flexibility in terms of footprint of the libraries used.

Signed-off-by: Franz Auernigg <f.auernigg@commend.com>