Signed-off-by: Simon McNair <simonmcnair@hotmail.com>
This is my first ever commit, please be gentle.
This is not my own work.  This is taken from https://github.com/SensorsIot/IOTstack/issues/415

This is added as per authorisation in this link https://github.com/SensorsIot/IOTstack/issues/415#issuecomment-932667190

"For the benefit of anyone reading this issue who would like to take the Mosquitto iotstack_healthcheck.sh script and either use it as-is or improve it and then propose a PR for Eclipse-Mosquitto, please go right ahead and do that"