support colon in username in pwfile via double colon
There are smart IoT devices that use the string represeantation of the MAC
address as MQTT user name, i.e., a user name containing colon characters.
In basic setups of Mosquitto with authentication based on pwfile, user names
could not contain colon characters, because the colon is used as a delimiter.

This commit adds supports for colon characters in user names in pwfile, by
interpreting the last colon in the pwline as the delimiter and the part in
font of it as user name. This works because the hased password never contains
a colon.

This approach has been suggested by Rob Swindell as response to PR#2619, which
suggested a more complicated approach.

Signed-off-by: Stefan Schuermans <stefan@schuermans.info>