support colon in username in pwfile via double colon
There are smart IoT devices that use the string represeantation of the MAC
address as MQTT user name, i.e., a user name containing colon characters.
In basic setups of Mosquitto with authentication based on pwfile, user names
could not contain colon characters, because the colon is used as a delimiter.

This commit adds supports for colon characters in user names in pwfile, by
interpreting double colons in the user name in pwfile as a single colon in
the actual user name.

Signed-off-by: Stefan Schuermans <stefan@schuermans.info>